requirements.txt
appdirs==1.4.2
beautifulsoup4==4.5.1
Django==1.10.4
django-anymail==0.9
django-bootstrap3==7.1.0
django-registration==2.2
django-recaptcha==1.3.1
google-api-python-client==1.5.5
httplib2==0.9.2
oauth2client==4.0.0
packaging==16.8
psycopg2==2.7.3
pyasn1==0.1.9
pyasn1-modules==0.0.8
pyparsing==2.1.10
requests==2.13.0
rsa==3.4.2
six==1.10.0
uritemplate==3.0.0
